Refactor mock fixtures in `test_manifest.py` to make them more widely available
